Initial commit of more OO version of Rex/Aux Kademlia support
