+++ lib/msf/core/auxiliary/mixins.rb
@@ -19,6 +19,7 @@ require 'msf/core/auxiliary/commandshell'
 require 'msf/core/auxiliary/login'
 require 'msf/core/auxiliary/rservices'
 require 'msf/core/auxiliary/cisco'
+require 'msf/core/auxiliary/kademlia'
 require 'msf/core/auxiliary/nmap'
 require 'msf/core/auxiliary/natpmp'
 require 'msf/core/auxiliary/iax2'
